requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
pandas==1.3.3
openpyxl==3.0.9
xmlrpc==0.9.1
requests==2.26.0
python-dotenv==0.19.0
